Dramatically improve pixel-setting performance by using unsafeFreeze
This change makes it so that when we freeze the array to make it
available to the drawing routine, we use unsafeFreeze instead of freeze.
This results in an O(1) conversion to an immutable array. It is safe in
this application because the drawing routine just needs an immutable
view of the array and will always only look at the array after it has
been updated anyway. For more info, see

http://hackage.haskell.org/package/array-0.5.2.0/docs/Data-Array-Unsafe.html#v:unsafeFreeze